+++ functions.php
@@ -3,7 +3,7 @@ namespace TSJIPPYTHEME;
 
 use Exception;
 
-define(__NAMESPACE__ .'\THEME', 'sim-theme');
+define(__NAMESPACE__ .'\THEME', 'tsjippy-theme');
 define(__NAMESPACE__ .'\THEME_PATH', str_replace('\\', '/', __DIR__));
 
 // We cannot use the sim theme
@@ -13,7 +13,7 @@ if(is_plugin_inactive('tsjippy-shared-functionality/tsjippy-shared-functionality
 
     switch_theme(array_values($themes)[0]['theme_root']);
     error_log("To use the tsjippy-theme you need to install the tsjippy-shared-functionality plugin");
-    //throw new Exception("To use the sim-theme you need to install the tsjippy-shared-functionality");
+    //throw new Exception("To use the tsjippy-theme you need to install the tsjippy-shared-functionality");
 }
 
 // composer
@@ -23,5 +23,3 @@ $files = glob(__DIR__  . '/php/*.php');
 foreach ($files as $file) {
     require_once($file);
 }
-
-//wp_enqueue_script('sim_theme_main_script', "$baseUrl/js/main.min.js", array(), wp_get_theme()->get('Version'), true);