+++ header-frontpage.php
@@ -1,5 +1,5 @@
 <?php
-namespace SIMTHEME;
+namespace TSJIPPYTHEME;
 
 /**
  * The template for displaying the frontpage header.