+++ js/main.js
@@ -1,3 +1,0 @@
-console.log('Main SIM-Theme js loaded');
-
-// footer always at the bottom